+++ app/src/main/java/com/ceaver/assin/assets/AssetRepository.kt
@@ -37,9 +37,9 @@ object AssetRepository {
                     Asset(
                             it.first.name,
                             it.first.symbol,
-                            it.second,
-                            it.first.priceBtc!!.times(it.second),
-                            it.first.priceUsd!!.times(it.second))
+                            it.second.toBigDecimal(),
+                            it.first.priceBtc!!.toBigDecimal().times(it.second.toBigDecimal()),
+                            it.first.priceUsd!!.toBigDecimal().times(it.second.toBigDecimal()))
                 }
     }
